added function for adding a time path to reduce redundant code
