+++ src/chain_actions/utils.rs
@@ -1,6 +1,6 @@
-use chrono::{DateTime, NaiveDateTime, Utc};
-use hdk::prelude::ExternResult;
-use hdk::time::sys_time;
+use chrono::{DateTime, NaiveDateTime, Utc, Datelike, Timelike};
+use holo_hash::EntryHash;
+use hdk::prelude::*;
 
 /// get the current UTC date time
 pub fn now_date_time() -> ExternResult<::chrono::DateTime<::chrono::Utc>> {
@@ -9,4 +9,20 @@ pub fn now_date_time() -> ExternResult<::chrono::DateTime<::chrono::Utc>> {
     let date: DateTime<Utc> =
         DateTime::from_utc(NaiveDateTime::from_timestamp(time.0, time.1), Utc);
     Ok(date)
+}
+
+pub fn add_current_time_path(base_component: String, entry_address: EntryHash) -> ExternResult<()>{
+    let date: DateTime<Utc> = now_date_time()?;
+
+    let time_path = crate::datetime_queries::utils::hour_path_from_date(
+        base_component,
+        date.year(),
+        date.month(),
+        date.day(),
+        date.hour(),
+    );
+
+    time_path.ensure()?;
+    create_link(time_path.hash()?, entry_address.clone(), ())?;
+    Ok(())
 }
